address review comments add documentation
