+++ constants.properties
@@ -1,4 +1,4 @@
-gradlePluginsVersion=4.0.32
+gradlePluginsVersion=4.0.33
 kotlinVersion=1.2.51
 # ***************************************************************#
 # When incrementing platformVersion make sure to update          #
