+++ docs/make-docsite.sh
@@ -11,12 +11,12 @@ else
     source virtualenv/Scripts/activate
 fi
 
-# TODO: The PDF rendering is pretty ugly and can be improved a lot.
-echo "Generating PDF document ..."
-make latexpdf
+## TODO: The PDF rendering is pretty ugly and can be improved a lot.
+#echo "Generating PDF document ..."
+#make latexpdf
 
 echo "Generating HTML pages ..."
 make html
-
-echo "Moving PDF file from $(eval echo $PWD/build/pdf/corda-developer-site.pdf) to $(eval echo $PWD/build/html/_static/corda-developer-site.pdf)"
-mv $PWD/build/latex/corda-developer-site.pdf $PWD/build/html/_static/corda-developer-site.pdf
+#
+#echo "Moving PDF file from $(eval echo $PWD/build/pdf/corda-developer-site.pdf) to $(eval echo $PWD/build/html/_static/corda-developer-site.pdf)"
+#mv $PWD/build/latex/corda-developer-site.pdf $PWD/build/html/_static/corda-developer-site.pdf
