--- samples/flow-override-demo/src/main/kotlin/net/corda/flows/overrides/SubResponder.kt
@@ -1,11 +1,0 @@
-package net.corda.flows.overrides
-
-import net.corda.core.flows.FlowSession
-import net.corda.core.flows.InitiatedBy
-
-@InitiatedBy(Initiator::class)
-class SubResponder(otherSideSession: FlowSession) : BaseResponder(otherSideSession) {
-    override fun getMessage(): String {
-        return "This is the sub responder"
-    }
-}
